add Makefile (#17)
We'd like to include uv next into the template. But then we have a lot
of places that need to check for the current dependency management tool
and run commands. A Makefile with targets makes it easier to
consistently run code checks also locally. We've tested this in other
projects and really like the simplicity of it. If the user doesn't like
the Makefile they can also just ignore it and paste the commands into
the pipeline.

### Changes

* add `Makefile` with 7 or 8 (with pre-commit) targets:
  * `help`: show all commands which comments as help
* `sync`: install or sync dependencies of the project including dev
dependencies
  * `install-ci`: install all dependencies required to run the pipeline
  * `mypy`: run mypy on the current codebase
  * `pylint`: run pylint on the current codebase
  * `ruff`: run ruff on the current codebase
  * `pre-commit` if used: run pre-commit on the current codebase
* `check`: run all three code checks (mypy, pylint and pre-commit or
ruff)
* update README and CONTRIBUTING accordingly
* use make commands in pipeline
* update tests to use new make commands