add option to use uv as dependency management tool (#18)
[`uv`](https://astral.sh/uv) is gaining in popularity due to its ease of
use. We'd like to test out if uv also works well for us and if we can
integrate it into our NiceGUI and RoSys setup. As a first step uv is
added to the options for a dependency management tool when we create new
projects with the template.

* update `use_poetry` question to `dependency_management` with three
options: `uv`, `poetry` or `pip`
* add three invisible questions to set the variables `use_uv`,
`use_poetry` and `use_pip` since which dependency management tool is
used is queried in the template quite extensively
* update pipeline to set up uv and use correct commands with it
* update `Makefile` commands for uv
* update `pyproject.toml`:
* if uv, pip or poetry without dynamic versioning is used, no build
system is added and we set `package-mode = false` for poetry, since most
project's will not published to PyPI
* if poetry with dynamic versioning is used, the corresponding build
system is used
* with this change we don't need to name our module the same as our
project, before this would have led to an error when uv or poetry would
look for a project to install and fine none with the correct name
* rename `requirements.txt`s to use check `if use_pip` instead of `if
not use_poetry`
* add uv as an option for running code checkers, update used python
versions
* update `README.md` of this repo and of the template accordingly, add
information about how to update a copied project

---------

Co-authored-by: Falko Schindler <mail@falkoschindler.de>